implement gpu i32 matrix multiplication

+++ leafeon-gpu/src/ops/mult.wgsl
@@ -1,7 +1,7 @@
 // Define the input matrix and result buffers
-@group(0) @binding(0) var<storage, read> A: array<f32>;  // Input matrix A (n x m)
-@group(0) @binding(1) var<storage, read> B: array<f32>;  // Input matrix B (m x p)
-@group(0) @binding(2) var<storage, read_write> C: array<f32>;  // Output matrix C (n x p)
+@group(0) @binding(0) var<storage, read> A: array<i32>;  // Input matrix A (n x m)
+@group(0) @binding(1) var<storage, read> B: array<i32>;  // Input matrix B (m x p)
+@group(0) @binding(2) var<storage, read_write> C: array<i32>;  // Output matrix C (n x p)
 
 // Matrix dimensions passed from the host side
 @group(0) @binding(3) var<uniform> matrix_dims: vec3<u32>;  // n, m, p
@@ -19,7 +19,7 @@ fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
 
     // Check if the indices are within bounds
     if (row < n) && (col < p) {
-        var sum: f32 = 0.0;
+        var sum: i32 = 0;
 
        // Perform the dot product of row from A and column from B
         for (var k = 0u; k < m; k = k + 1u) {
@@ -29,5 +29,5 @@ fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
         // Store the result in C
         C[row * p + col] = sum;
         //C[row * m + col] = sum;
-   }
+    }
 }
